+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_grpc_web_pb.js
@@ -1048,5 +1048,66 @@ proto.api_container_api.ApiContainerServicePromiseClient.prototype.uploadFilesAr
 };
 
 
+/**
+ * @const
+ * @type {!grpc.web.MethodDescriptor<
+ *   !proto.api_container_api.DownloadFilesArtifactArgs,
+ *   !proto.api_container_api.DownloadFilesArtifactResponse>}
+ */
+const methodDescriptor_ApiContainerService_DownloadFilesArtifact = new grpc.web.MethodDescriptor(
+  '/api_container_api.ApiContainerService/DownloadFilesArtifact',
+  grpc.web.MethodType.UNARY,
+  proto.api_container_api.DownloadFilesArtifactArgs,
+  proto.api_container_api.DownloadFilesArtifactResponse,
+  /**
+   * @param {!proto.api_container_api.DownloadFilesArtifactArgs} request
+   * @return {!Uint8Array}
+   */
+  function(request) {
+    return request.serializeBinary();
+  },
+  proto.api_container_api.DownloadFilesArtifactResponse.deserializeBinary
+);
+
+
+/**
+ * @param {!proto.api_container_api.DownloadFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>} metadata User defined
+ *     call metadata
+ * @param {function(?grpc.web.RpcError, ?proto.api_container_api.DownloadFilesArtifactResponse)}
+ *     callback The callback function(error, response)
+ * @return {!grpc.web.ClientReadableStream<!proto.api_container_api.DownloadFilesArtifactResponse>|undefined}
+ *     The XHR Node Readable Stream
+ */
+proto.api_container_api.ApiContainerServiceClient.prototype.downloadFilesArtifact =
+    function(request, metadata, callback) {
+  return this.client_.rpcCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/DownloadFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_DownloadFilesArtifact,
+      callback);
+};
+
+
+/**
+ * @param {!proto.api_container_api.DownloadFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>=} metadata User defined
+ *     call metadata
+ * @return {!Promise<!proto.api_container_api.DownloadFilesArtifactResponse>}
+ *     Promise that resolves to the response
+ */
+proto.api_container_api.ApiContainerServicePromiseClient.prototype.downloadFilesArtifact =
+    function(request, metadata) {
+  return this.client_.unaryCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/DownloadFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_DownloadFilesArtifact);
+};
+
+
 module.exports = proto.api_container_api;