+++ core/docs/changelog.md
@@ -1,8 +1,8 @@
 # TBD
-
-### Changes
+### Features
 * Added new backend file storage methods for future volume support.
 * Added UploadFilesArtifact command to API Container.
+* Add a `DownloadFilesArtifact` endpoint to the API container for downloading files artifacts from the web
 
 # 1.43.0
 ### Breaking Changes